fix(ui): align auth-cold picker CI contracts
