+++ ui/src/components/web-awesome-migration.node.test.ts
@@ -53,6 +53,7 @@ describe("Web Awesome control ownership", () => {
       "components/command-palette.ts",
       "pages/chat/components/chat-composer-skill-menu.ts",
       "pages/chat/components/chat-composer-slash-menu.ts",
+      "pages/chat/components/chat-model-picker-options.ts",
       "pages/chat/components/chat-model-picker.ts",
     ]);
   });
